Move specialized method groups to mixin classes (#301)
